Remove goofy pip handler
